feat: Error handling w/o args
